New method NexT.utils.activateSidebarPanel
* See: https://github.com/next-theme/hexo-theme-next/issues/368
