+++ source/js/utils.js
@@ -305,6 +305,34 @@ NexT.utils = {
     }
   },
 
+  activateSidebarPanel: function(index) {
+    const duration = 200;
+    const sidebar = document.querySelector('.sidebar-inner');
+    const panel = document.querySelector('.sidebar-panel-container');
+    const activeClassName = ['sidebar-toc-active', 'sidebar-overview-active'];
+
+    if (sidebar.classList.contains(activeClassName[index])) return;
+
+    window.anime({
+      duration,
+      targets   : panel,
+      easing    : 'linear',
+      opacity   : 0,
+      translateY: [0, -20],
+      complete  : () => {
+        // Prevent adding TOC to Overview if Overview was selected when close & open sidebar.
+        sidebar.classList.replace(activeClassName[1 - index], activeClassName[index]);
+        window.anime({
+          duration,
+          targets   : panel,
+          easing    : 'linear',
+          opacity   : [0, 1],
+          translateY: [-20, 0]
+        });
+      }
+    });
+  },
+
   getScript: function(src, options = {}, legacyCondition) {
     if (typeof options === 'function') {
       return this.getScript(src, {
